- Moved vfs server into folder fs - Added plain fs servers for fat32, ext4 and dev below fs - Moved mailbox and mmio server into folder io - Added plain sdcard storage and random server - Adjusted paths at boot server after move - Added link of libconfini to boot server
