+++ bolthur/server/Makefile.am
@@ -1,4 +1,4 @@
 
 ACLOCAL_AMFLAGS = -I ../../build-aux/m4
 
-SUBDIRS = boot console platform terminal vfs
+SUBDIRS = boot console fs platform terminal
