+++ bolthur/server/fs/Makefile.am
@@ -1,0 +1,2 @@
+
+SUBDIRS = dev ext4 fat32 vfs
