+++ bolthur/server/fs/dev/Makefile.am
@@ -1,0 +1,7 @@
+
+AM_CFLAGS = -DPROGRAM_NAME=\"fs/dev\"
+
+bin_PROGRAMS = dev
+dev_SOURCES = \
+  main.c
+dev_LDFLAGS = -all-static --static
