+++ bolthur/server/fs/ext4/Makefile.am
@@ -1,0 +1,7 @@
+
+AM_CFLAGS = -DPROGRAM_NAME=\"fs/ext4\"
+
+bin_PROGRAMS = ext4
+ext4_SOURCES = \
+  main.c
+ext4_LDFLAGS = -all-static --static
