+++ bolthur/server/fs/fat32/Makefile.am
@@ -1,0 +1,7 @@
+
+AM_CFLAGS = -DPROGRAM_NAME=\"fs/fat32\"
+
+bin_PROGRAMS = fat32
+fat32_SOURCES = \
+  main.c
+fat32_LDFLAGS = -all-static --static
