+++ bolthur/server/fs/vfs/rpc/remove.c
