+++ bolthur/server/fs/vfs/rpc/stat.c
