+++ bolthur/server/platform/raspi/io/Makefile.am
@@ -1,0 +1,2 @@
+
+SUBDIRS = mailbox mmio
