+++ bolthur/server/platform/raspi/io/mmio/Makefile.am
@@ -1,5 +1,5 @@
 
-AM_CFLAGS = -DPROGRAM_NAME=\"mmio\"
+AM_CFLAGS = -DPROGRAM_NAME=\"io/mmio\"
 
 bin_PROGRAMS = mmio
 mmio_SOURCES = \
