+++ bolthur/server/platform/raspi/random/Makefile.am
@@ -1,0 +1,7 @@
+
+AM_CFLAGS = -DPROGRAM_NAME=\"random\"
+
+bin_PROGRAMS = random
+random_SOURCES = \
+  main.c
+random_LDFLAGS = -all-static --static
