+++ bolthur/server/platform/raspi/storage/Makefile.am
@@ -1,0 +1,2 @@
+
+SUBDIRS = sdcard
