+++ bolthur/server/platform/raspi/storage/sdcard/Makefile.am
@@ -1,0 +1,7 @@
+
+AM_CFLAGS = -DPROGRAM_NAME=\"storage/sdcard\"
+
+bin_PROGRAMS = sdcard
+sdcard_SOURCES = \
+  main.c
+sdcard_LDFLAGS = -all-static --static
